Conflicts:
	MPC.Web/Areas/Stores/Scripts/Stores.viewModel.js

+++ Models/DomainModels/TemplateVariable.cs
@@ -6,7 +6,7 @@
     public class TemplateVariable
     {
         public long ProductVariableId { get; set; }
-        public int? TemplateId { get; set; }
-        public int? VariableId { get; set; }
+        public long? TemplateId { get; set; }
+        public long? VariableId { get; set; }
     }
 }
